refactor: board 도메인 패키징
